added state range conversion to float to insure values are never divided as ints

+++ representation.py
@@ -298,7 +298,7 @@ class Tiling(object):
         else:
             ntiles = np.array(ntiles)
             
-        self.state_range = [state_range[0][input_index].copy(), state_range[1][input_index].copy()]
+        self.state_range = [state_range[0][input_index].copy().astype(float), state_range[1][input_index].copy().astype(float)]
         self.state_range[0] -= (self.state_range[1]-self.state_range[0])/(ntiles-1)
 
         self.offset = offset
